Renames find_citations.py to find.py.
